+++ src/queries/index.ts
@@ -1,4 +1,4 @@
 export * from './productsQueries';
 export * from './cartQueries';
-export * from './orderQueries';
+export * from './ordersQueries';
 export * from './paymentsQueries';
